+++ libcoffee/molecule/rdkit/mol.py
@@ -23,7 +23,7 @@ class Mol(MolBase):
     @property
     def isotopes(self) -> npt.NDArray[np.int32]:
         # a.GetIsotope() should return int but typing says return Any
-        return np.array([a.GetIsotope() for a in self._atoms], dtype=np.int32)  # type: ignore  # Argument is not needed for GetIsotope()
+        return np.array([a.GetIsotope() for a in self._atoms], dtype=np.int32)
 
     @isotopes.setter
     def isotopes(self, isotopes: npt.NDArray[np.int32]) -> None:
@@ -37,7 +37,7 @@ class Mol(MolBase):
         return self._mol.GetProp("_Name")  # type: ignore[no-any-return]
 
     @property
-    def heavy_atom_indices(self) -> npt.NDArray[np.int32]:
+    def heavy_atom_indices(self) -> npt.NDArray[np.intp]:
         atomic_nums = [a.GetAtomicNum() for a in self._atoms]
         return np.where(np.array(atomic_nums) > 1)[0]
 
@@ -57,13 +57,13 @@ class Mol(MolBase):
             coords = coords[self.heavy_atom_indices]
         return coords
 
-    def extract_submol(self, atom_idxs: npt.NDArray[np.int32]) -> "Mol":
+    def extract_submol(self, atom_idxs: npt.NDArray[np.intp]) -> "Mol":
         rw_mol = Chem.RWMol(self.raw_mol)
         idx_remove_atoms = set(range(self.raw_mol.GetNumAtoms())) - set(atom_idxs)
         atomidxs = sorted(idx_remove_atoms)[::-1]
         for idx in atomidxs:
             rw_mol.RemoveAtom(idx)
-        return Mol(rw_mol.GetMol())  # type: ignore  # Argument of GetMol() is not needed
+        return Mol(rw_mol.GetMol())
 
     def merge(self, mol: "Mol", aps: tuple[int, int] | None = None) -> "Mol":
         raise NotImplementedError
